make max depth 255 instead of 63
passed LTC:
Elo   | 1.68 +- 2.61 (95%)
SPRT  | 40.0+0.40s Threads=1 Hash=128MB
LLR   | 2.99 (-2.94, 2.94) [-3.50, 0.50]
Games | N: 19416 W: 4645 L: 4551 D: 10220
Penta | [110, 2281, 4841, 2357, 119]
https://nonlinear.eu.pythonanywhere.com/test/1010/
passed STC:
Elo   | 5.52 +- 4.41 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=32MB
LLR   | 2.95 (-2.94, 2.94) [-3.50, 0.50]
Games | N: 8124 W: 2130 L: 2001 D: 3993
Penta | [86, 947, 1895, 1020, 114]
https://nonlinear.eu.pythonanywhere.com/test/1009/
bench: 2539706

+++ src/core.cpp
@@ -695,7 +695,7 @@ int Engine::negamax(int depth, int alpha, int beta, Stack* ss, bool cutnode){
     assert(is_valid(max_value));
 
     transposition_table.store(zobrist_hash, to_tt(max_value, ply), uncorrected_static_eval, depth, best_move,
-        node_type, static_cast<uint8_t>(pos.fullMoveNumber()), pv);
+        node_type, pos.fullMoveNumber(), pv);
 
     return max_value;
 }
@@ -792,7 +792,7 @@ int Engine::qsearch(int alpha, int beta, int depth, Stack* ss){
         if (stand_pat >= beta){
             if (!is_hit)
                 transposition_table.store(zobrist_hash, to_tt(stand_pat, ply), uncorrected_static_eval,
-                    DEPTH_QSEARCH, Move::NO_MOVE, TFlag::EXACT, static_cast<uint8_t>(pos.fullMoveNumber()), pv);
+                    DEPTH_QSEARCH, Move::NO_MOVE, TFlag::EXACT, pos.fullMoveNumber(), pv);
             return stand_pat;
         }
 
@@ -878,7 +878,7 @@ int Engine::qsearch(int alpha, int beta, int depth, Stack* ss){
         }
 
         transposition_table.store(zobrist_hash, to_tt(stand_pat, ply), NO_VALUE, DEPTH_QSEARCH,
-            Move::NO_MOVE, TFlag::EXACT, static_cast<uint8_t>(pos.fullMoveNumber()), pv);
+            Move::NO_MOVE, TFlag::EXACT, pos.fullMoveNumber(), pv);
         return stand_pat;
     }
 
@@ -893,7 +893,7 @@ int Engine::qsearch(int alpha, int beta, int depth, Stack* ss){
         transposition_table.store(zobrist_hash, to_tt(max_value, ply),
             uncorrected_static_eval, DEPTH_QSEARCH, best_move,
             max_value >= beta ? TFlag::LOWER_BOUND : TFlag::UPPER_BOUND,
-            static_cast<uint8_t>(pos.fullMoveNumber()), pv);
+            pos.fullMoveNumber(), pv);
 
     return max_value;
 }
